Run 2to3 against Python code
The 2to3 utility is an officially-supported script to automatically convert
Python 2 code to Python 3.  It's not a complete fix by any means, but it
handles most basic syntax transformations and similar.

This commit is the result of running 2to3 against every Python file in the
gpMgmt directory, so it's quite large and fairly scattershot.  Manual updates
to any code that 2to3 can't handle will come in later commits.