Remove serializable argument to GetSnapshotData()
In `GetSnapshotData()` serializable was passed to make sure first
snapshot is setting the `MyProc->xmin` and also Assert for the
same. Instead of passing the flag, use the
`TransactionIdIsValid(MyProc->xmin)` or not to set the same. This is
similar to how its done in latest version of GPDB6+ and postgres post
commit 5da9da71c44f27ba48fdad08ef263bf70e43e689.

The reason to make the change is for scanning
gp_segment_configuration, need to acquire snapshot to scan the
catalog. Don't wish to touch variable `SerializableSnapshot` for
acquiring this snapshot, since there are places in code which check
for this variable to be NULL to detect if any query was run
before. For example `set transaction isolation level`. Hence, reading
gp_segment_configuration shouldn't be treated as running any query,
before acquiring snapshot. So, need way to get the latest snaphot and
then later regular query can still call GetTransactionSnapshot() and
still acquire serializable snapshot that time.