Convey if QD encountered delim or not to QE
This is necessary to differentiate between empty/missing next
attribute from empty next attribute. QD encountering delim, means next
attribute is empty and not really missing. QE needs to know this
information, as it has no access to find this info, since the location
QD ask QE to start reading is after delim if present or end of line.