Update autovacuum test
After catalog table autovacuums are enabled across all databases,
autovacuum test must be updated to be database agnostic. Issue is that
autovacuum always operates on the oldest database and since we only
allow autovacuum to operate on catalog tables we are not guaranteed to
ever autovacuum template0 during the course of this test. This is
because an old user table on another database would block autovacuum
from ever choosing template0.

Now the test will check age of a catalog table on the oldest database
and verify that after exhausting 200 million transactions the table is
autovacuumed and the age remains less than 200 million.