Avoid possibility of out-of-bound write for neededColumnContextWalker
neededColumnContextWalker() is called to scan through VARs for targetlist,
quals, etc.. It should only look at VARS for the table being scanned and
avoid all other VARS. Currently, we are not aware of any plans which can
produce situation where neededColumnContextWalker() will encounter some
other VARs. But for GPDB5, we get OUTER vars here if Index scan is right
tree for NestedLoop join. Hence, seems better to have the protective
code to not write out-of-bound.

Adds test to cover the scenario as well which is missing currently.

Co-authored-by: Alexandra Wang <lewang@pivotal.io>